Reorganized the file, completed the update functionality, and changed the layout
